notes/requirements.txt
Django==2.1.1
django-rest-auth==0.9.3
django-templated-mail==1.1.1
django-widget-tweaks==1.4.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
mysqlclient==1.3.13
python-decouple==3.1
pytz==2018.5
six==1.11.0
